% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Priors.R
\name{create_dichotomous_prior}
\alias{create_dichotomous_prior}
\title{create_dichotomous_prior  Given priorlist, a model, 
       and a distribution. Create a prior for a given analysis.}
\usage{
create_dichotomous_prior(prior, model)
}
\arguments{
\item{prior}{First Prior}

\item{model}{Model to be used should be one of"hill","gamma","logistic","log-logistic","log-probit","multistage", "probit", "qlinear", or "weibull"}
}
\value{
new BMDprior list that can be used in a dichotomous fit.
}
\description{
create_dichotomous_prior  Given priorlist, a model, 
       and a distribution. Create a prior for a given analysis.
}
\examples{
plist<- create_prior_list(normprior(0,0.1,-100,100), # a
                          lnormprior(1,0.2,0,18))
 
 power_normal <- create_dichotomous_prior(plist,"logistic") 

}
