% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_classes.R
\name{create_prior_list}
\alias{create_prior_list}
\title{create_prior_lists .. Given priors 
       created using the ToxicR prior functions, create a list of priors
       for a model.}
\usage{
create_prior_list(x1, x2, ...)
}
\arguments{
\item{x1}{First Prior}

\item{x2}{Second Prior}

\item{...}{Aditional arguments}
}
\value{
new BMDprior list. This object is essentailly a matrix where each
row is  an element defined by a prior object (e.g., normprior or lnormprior).
}
\description{
create_prior_lists .. Given priors 
       created using the ToxicR prior functions, create a list of priors
       for a model.
}
\examples{
plist<- create_prior_list(normprior(0,0.1,-100,100), # a
                          normprior(0,1,  -1e2,1e2),     # b
                          lnormprior(1,0.2,0,18),  #k
                          normprior(0,1,-18,18))

}
