% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{data}
\name{dichotomousDR}
\alias{dichotomousDR}
\title{733 unique dose-response datasets}
\format{
A data frame with 2727 rows and 11 variables:
\describe{
  \item{ID}{-The study ID in the database.}
  \item{chemical}{-Name of the Chemical in the study.}
  \item{data.source}{-Source of the risk assessment data.}
  \item{CASRN}{-Chemical's CASRN}
  \item{dose}{-Dose spacing of the study using the original study.}
  \item{r.dose}{-Doses of the experiment relative to 1 being the maximum dose tested.}
  \item{n}{-Number of animals on test.}
  \item{obs}{-Number of adverse events.}
  \item{organ}{-Organ impacted.}
  \item{effect}{-Type of adverse effect.}
  \item{study.source}{-Publication related to the experiment.}
}
More information at: \doi{10.1111/risa.13218}
}
\usage{
dichotomousDR
}
\description{
A dataset containing 733 dichotomous dose-response studies that were involved in 
regulatory risk assessment.
}
\keyword{datasets}
