% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_classes.R
\name{gammaprior}
\alias{gammaprior}
\title{gammaprior - create a Gamma prior.}
\usage{
gammaprior(mean = 1, sd = 1, lb = 0, ub = 100)
}
\arguments{
\item{mean}{mean of the prior distribution.}

\item{sd}{sd of the prior distribution.}

\item{lb}{lower bound on the distribution. Necessary for the optimization algorithms, 
must be >= 0}

\item{ub}{Upper bound on the distribution. Necessary for the optimization algorithms,
To make sure it is a fully Gamma prior, make ub large relative to the mean/sd.}
}
\value{
a normal prior model object.  This object is essentially a vector with
the first element as 4 (for Gamma), the second element the mean, the third
element the standard deviation, the fourth and fifth elements the lower and upper bounds, respectively.
}
\description{
Specify a Gamma prior for a ToxicR Bayesian model fit
}
\examples{
# Gamma Prior with mean 1,sd 1, and lower bound 0.1
gammaprior(mean = 1, sd = 1, lb = 0.1, ub = 1e4)
}
