% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Priors.R
\name{ma_build_priors}
\alias{ma_build_priors}
\title{ma_build_priors - Build a list of default priors for model averaging}
\usage{
ma_build_priors(
  Y = NA,
  model_list = NA,
  distribution_list = NA,
  ma_type = "EFSA"
)
}
\arguments{
\item{Y}{Response matrix (optional) but preferred to fix variance prior}

\item{model_list}{Vector of models to use. See \code{\link{single_continuous_fit}} for details on the models. Overrides ma_type}

\item{distribution_list}{Vector of corresponding underlying distribution to use ("normal", "normal-ncv", "lognormal"). Defaults to normal-ncv}

\item{ma_type}{Models to average over ("original", "EFSA", or "all"). All corresponds to all implemented models; 
EFSA and original are described in \code{\link{ma_continuous_fit}} under the EFSA parameter.}
}
\value{
A list of configurations for the single models (priors, model type). Can be fed as the model_list to ma_continuous_fit
}
\description{
Build a list of default priors for model averaging
}
\examples{
#get EFSA model priors
prior_list <- ma_build_priors(ma_type = "EFSA")

}
