% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging_fits.R
\name{ma_continuous_fit}
\alias{ma_continuous_fit}
\title{ma_continuous_fit - Fit a model averaged continuous BMD model.}
\usage{
ma_continuous_fit(
  D,
  Y,
  model_list = NA,
  fit_type = "laplace",
  BMD_TYPE = "sd",
  BMR = 0.1,
  point_p = 0.01,
  alpha = 0.05,
  samples = 21000,
  burnin = 1000
)
}
\arguments{
\item{D}{doses matrix}

\item{Y}{response matrix}

\item{model_list}{a list of configurations for the single models (priors, model type)}

\item{fit_type}{the method used to fit (laplace, mle, or mcmc)}
}
\description{
Fit a model averaged continuous BMD model.
}
\examples{
Hill.p <- rbind(c(481,-250.3,70,3.3),
                c(481,-250.3,40,1.3),
                c(481,-250.2,15,1.1),
                c(481,-250.3,50,4) ,
                c(10.58,9.7,70,3.5),
                c(10.58,9.7,25,3),
                c(10.58,9.7,15,2),
                c(10.58,9.7,50,4))
hill <- data.frame(a=Hill.p[,1],b=Hill.p[,2],c=Hill.p[,3],d=Hill.p[,4])
doses <- rep(c(0,6.25,12.5,25,50,100),each=10)
dosesq <- rep(c(0,6.25,12.5,25,50,100),each=30)
mean <- cont_hill_f(as.numeric(hill[2,]),doses)
y <- rinvgauss(length(mean),mean,18528.14)
model <- ma_continuous_fit(doses, y, fit_type = "laplace", BMD_TYPE = 'sd', BMR = 1)

}
