% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAdensity_plot.R
\name{MAdensity_plot}
\alias{MAdensity_plot}
\title{MAdensity_plot - Create a density plot from a model averaged model.}
\usage{
MAdensity_plot(A)
}
\arguments{
\item{A}{the model averaged model to plot}
}
\value{
Returns a \code{ggplot2} graphics object.
}
\description{
Create a density plot from a model averaged model fit with MCMC.
}
\examples{
\donttest{
doses  <- cbind(c(0,25,50,100,200))
y <- cbind(c(6,5.2,2.4,1.1,0.75),
                c(20,20,19,20,20),
                c(1.2,1.1,0.81,0.74,0.66))
model <- ma_continuous_fit(doses,y,
                        fit_type = "mcmc",BMD_TYPE = 'sd',BMR = 1)
MAdensity_plot(model)
}
}
