% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{data}
\name{ntp_599_female}
\alias{ntp_599_female}
\title{Long term Thyroid Adenoma data from NTP Report 599}
\format{
A data frame with 200 rows and 4 variables:
\describe{
  \item{treatment}{-The dose group for the observation.}
  \item{days_on_study}{-Number of days on the study 730 is the max.}
  \item{adenoma}{- Thyroid Adenoma (Yes/No) (1/0).}
  \item{dose}{-The dose in mg/L}
}
For more information see: \doi{10.22427/NTP-DATA-TR-599}
}
\usage{
ntp_599_female
}
\description{
This dataset contains Thyroid Adenoma data  for  
female rats for the technical report TR-599: Sodium Tungstate Dihydrate.
}
\keyword{datasets}
