% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{data}
\name{ntp_599_hemotology}
\alias{ntp_599_hemotology}
\title{Clinical Chemistry data from NTP Report 599}
\format{
A data frame with 200 rows and 4 variables:
\describe{
  \item{concentration}{-The dose group for the observation.}
  \item{sex}{- Male/Female.}
  \item{response}{- Response variable}
  \item{response_type}{- The type of response measured}
}
For more information see: \doi{10.22427/NTP-DATA-TR-599}
}
\usage{
ntp_599_hemotology
}
\description{
This dataset contains clinical chemistry data  for  
all rats in the short term 90-day study.
}
\keyword{datasets}
