% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{data}
\name{ntp_weight_data}
\alias{ntp_weight_data}
\title{Short term terminal body-weight data from NTP Report 599}
\format{
A data frame with 120 rows and 4 variables:
\describe{
  \item{Dose_Group}{-The dose group for the observation.}
  \item{dose}{-The dose in mg/L }
  \item{sex}{-Animal's Sex}
  \item{weight}{-Terminal body-weight}
}
For more information see: \doi{10.22427/NTP-DATA-TR-599}
}
\usage{
ntp_weight_data
}
\description{
This dataset contains terminal body-weight data for male and 
female rats for the technical report TR-599: Sodium Tungstate Dihydrate.
}
\keyword{datasets}
