% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomous_wrappers.R
\name{single_dichotomous_fit}
\alias{single_dichotomous_fit}
\title{Fit a single dichotomous dose-response model to data.}
\usage{
single_dichotomous_fit(
  D,
  Y,
  N,
  model_type,
  fit_type = "laplace",
  prior = "default",
  BMR = 0.1,
  alpha = 0.05,
  degree = 2,
  samples = 21000,
  burnin = 1000
)
}
\arguments{
\item{D}{A numeric vector of doses. Here doses are $\geq 0.$}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Fit a single dichotomous dose-response model to data.
}
\examples{
add(1, 1)
add(10, 1)
}
