% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ToxicR.R
\name{update_ToxicR}
\alias{update_ToxicR}
\title{update_ToxicR - downloads and tries to install the latest version of the package from github
Updates the current package 'ToxicR' by installing the
most recent version of the package from GitHub
This function requires installing Package 'remotes' v2.4.2
(or possibly a higher version) by Csardi et al. (2021),
<https://cran.r-project.org/package=remotes>}
\usage{
update_ToxicR(force = TRUE, upgrade_other_pkg = FALSE, confirm = TRUE)
}
\arguments{
\item{force}{logical. If \code{force = TRUE}, force installing the
update. If \code{force = FALSE}, do not force installing the update.
By default, \code{force = TRUE}.}

\item{upgrade_other_pkg}{input for the \code{upgrade} argument to
be passed on to \code{remotes::install_github}.
One of "default", "ask", "always", "never", TRUE, or FALSE.
"default" respects the value of the R_REMOTES_UPGRADE environment
variable if set, and falls back to "ask" if unset.
"ask" prompts the user for which out of date packages to upgrade.
For non-interactive sessions "ask" is equivalent to "always".
TRUE and FALSE correspond to "always" and "never" respectively.
By default, \code{upgrade_other_pkg = FALSE}.}

\item{confirm}{logical. If \code{confirm = TRUE}, the user will
need to confirm the update. If \code{confirm = FALSE}, the confirmation
step will be skipped. By default, \code{confirm = TRUE}.}
}
\value{
there will be no output from this function. Rather, executing
this function will update the current 'ToxicR' package by installing
the most recent version of the package from GitHub.
}
\description{
Update the package 'ToxicR'
}
\examples{
\dontrun{
if (interactive()) {update_ToxicR()}
}

}
