% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleveland_plot.R
\name{cleveland_plot}
\alias{cleveland_plot}
\title{cleveland_plot - Create a Cleveland plot from a model averaged model.}
\usage{
cleveland_plot(A, ...)
}
\arguments{
\item{A}{the model averaged model to plot}
}
\description{
Create a Cleveland plot from a model averaged model.
}
\examples{

...
model = ma_dichotomous_fit(D,Y,N)
cleveland_plot(model)

}
