% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging_fits.R
\name{ma_dichotomous_fit}
\alias{ma_dichotomous_fit}
\title{ma_dichotomous_fit - Fit a model averaged dichotomous BMD model.}
\usage{
ma_dichotomous_fit(
  D,
  Y,
  N,
  model_list = integer(0),
  fit_type = "laplace",
  BMD_TYPE = "extra",
  BMR = 0.1,
  point_p = 0.01,
  alpha = 0.05,
  samples = 21000,
  burnin = 1000
)
}
\arguments{
\item{D}{doses matrix}

\item{Y}{response matrix}

\item{N}{number of replicates matrix}

\item{model_list}{a list of configurations for the single models (priors, model type)}

\item{fit_type}{the method used to fit (laplace, mle, or mcmc)}
}
\description{
Fit a model averaged dichotomous BMD model.
}
\examples{

mData <- matrix(c(0, 2,50,
                  1, 2,50,
                  3, 10, 50,
                  16, 18,50,
                  32, 18,50,
                  33, 17,50),nrow=6,ncol=3,byrow=T)
D <- mData[,1]
Y <- mData[,2]
N <- mData[,3]
model = ma_dichotomous_fit(D,Y,N)

}
