% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAdensity_plot.R
\name{MAdensity_plot}
\alias{MAdensity_plot}
\title{MAdensity_plot - Create a density plot from a model averaged model.}
\usage{
MAdensity_plot(A, ...)
}
\arguments{
\item{A}{the model averaged model to plot}
}
\description{
Create a density plot from a model averaged model fit with MCMC.
}
\examples{

...
model <- ma_continuous_fit(doses,y,model_list=model_list,
                        fit_type = "mcmc",BMD_TYPE = 'sd',BMR = 1)
MAdensity_plot(model)

}
