% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NTP.R
\docType{data}
\name{NTP::williams}
\alias{NTP$williams}
\title{Williams Trend test for}
\format{
An object of class \code{module} (inherits from \code{list}) of length 5.
}
\usage{
NTP$willams( ) 
}
\arguments{
\item{data}{A data frame with column names in the formula.}
}
\value{
The results of a Williams trend test for each level in \code{dose_name}.
}
\description{
Williams Trend test for
}
\examples{
add(1, 1)
add(10, 1)
}
\keyword{datasets}
