% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomous_wrappers.R
\name{single_dichotomous_fit}
\alias{single_dichotomous_fit}
\title{Fit a single dichotomous dose-response model to data.}
\usage{
single_dichotomous_fit(
  D,
  Y,
  N,
  model_type,
  fit_type = "laplace",
  prior = "default",
  BMR = 0.1,
  alpha = 0.05,
  degree = 2,
  samples = 21000,
  burnin = 1000
)
}
\arguments{
\item{D}{A numeric vector or matrix of doses.}

\item{Y}{A numeric vector or matrix of responses.}

\item{N}{A numeric vector or matrix of the number of replicates at a dose.}

\item{fit_type}{the method used to fit (laplace, mle, or mcmc)}

\item{degree}{the number of degrees of a polynomial model. Only used for polynomial models.}

\item{samples}{the number of samples to take (MCMC only)}

\item{burnin}{the number of burnin samples to take (MCMC only)}

\item{BRM}{This option specifies the benchmark response BMR. The BMR is defined in relation to the BMD calculation requested (see BMD).  By default, the "BMR = 0.1."}
}
\value{
a model object
}
\description{
Fit a single dichotomous dose-response model to data.
}
\examples{
mData <- matrix(c(0, 2,50,
                  1, 2,50,
                  3, 10, 50,
                  16, 18,50,
                  32, 18,50,
                  33, 17,50),nrow=6,ncol=3,byrow=T)
D <- mData[,1]
Y <- mData[,2]
N <- mData[,3]
model = single_dichotomous_fit(D, Y, N, model_type = "hill", fit_type = "laplace")

}
